feat(serializer): Use placeholder classnames (#12)
Closes #11

This replaces the glamor generated classnames in the output with placeholder "glamor-*" classnames. This improves diffs as they are not polluted by the change in classnames.

BREAKING CHANGE: Your snapshots will all break and will need to be updated. That's all though :smile: and they'll be much easier to read!